+++ R/dotprops.R
@@ -285,8 +285,9 @@ all.equal.dotprops<-function(target, current, check.attributes=FALSE,
 #'   \code{alpha}
 #' @param ... Additional arguments passed to \code{points3d} and/or 
 #'   \code{segments3d}
+#' @inheritParams plot3d.neuronlist
 #' @return invisible list of results of rgl plotting commands
-#' @method plot3d dotprops
+#' 
 #' @export
 #' @seealso \code{\link{dotprops}, \link[rgl]{plot3d}, \link[rgl]{points3d}, 
 #'   \link[rgl]{segments3d}}
@@ -294,20 +295,36 @@ all.equal.dotprops<-function(target, current, check.attributes=FALSE,
 #' \donttest{
 #' open3d()
 #' plot3d(kcs20[[1]])
-#' clear3d()
+#' nclear3d()
 #' plot3d(kcs20[[1]],col='red')
-#' clear3d()
+#' nclear3d()
 #' plot3d(kcs20[[1]],col='red',lwd=2)
 #' plot3d(kcs20[[2]],col='green',lwd=2)
 #' }
-plot3d.dotprops<-function(x, scalevecs=1.0, alpharange=NULL, color='black', 
+plot3d.dotprops<-function(x, scalevecs=1.0, plotengine = getOption('nat.plotengine'),
+                          alpharange=NULL, color='black',
                           PlotPoints=FALSE, PlotVectors=TRUE, UseAlpha=FALSE, ...){
   # rgl's generic plot3d will dispatch on this
   if (!is.null(alpharange))
     x=subset(x,x$alpha<=alpharange[2] & x$alpha>=alpharange[1])
+  plotengine <- check_plotengine(plotengine)
   rlist=list()
+  if (plotengine == 'plotly') {
+    psh <- openplotlyscene()$plotlyscenehandle
+    params=list(...)
+    opacity <- if("alpha" %in% names(params)) params$alpha else 1
+  }
+  
   if(PlotPoints){
-    rlist$points=points3d(x$points, color=color, ...)
+    if (plotengine == 'rgl'){
+      rlist$points=points3d(x$points, color=color, ...)
+    } else {
+      plotdata <- as.data.frame(x$points)
+      psh <- psh %>% 
+        plotly::add_trace(data = plotdata, x = ~X, y = ~Y , z = ~Z, 
+        hoverinfo = "none",type = 'scatter3d', mode = 'markers',
+        opacity = opacity, marker=list(color = color, size = 3))
+    }
   }
     
   if(PlotVectors){
@@ -319,9 +336,33 @@ plot3d.dotprops<-function(x, scalevecs=1.0, alpharange=NULL, color='black',
     starts=x$points-halfvect
     stops=x$points+halfvect
     interleaved=matrix(t(cbind(starts,stops)),ncol=3,byrow=T)
-    rlist$segments=segments3d(interleaved, color=color, ...)
+    if (plotengine == 'rgl'){
+        rlist$segments=segments3d(interleaved, color=color, ...)
+    } else {
+      tempdata <- interleaved
+      tempseglist <- list()
+      for (tempidx in seq(1,nrow(tempdata)/2)) {
+        tempseglist[[tempidx]] <- c(1,2)+2*(tempidx-1)
+      }
+      
+      tempdata<-do.call(rbind,sapply(tempseglist,function(s) {rbind(tempdata[s,],NA)},simplify=FALSE))
+      
+      plotdata <- as.data.frame(tempdata)
+      names(plotdata) <- c('X','Y','Z')
+      psh <- psh %>% 
+        plotly::add_trace(data = plotdata, x = ~X, y = ~Y , z = ~Z, 
+        hoverinfo = "none", type = 'scatter3d', mode = 'lines',
+        opacity = opacity, line=list(color = color, width = 4))
+    }
   }
-  invisible(rlist)
+  if (plotengine == 'rgl'){
+    invisible(rlist)
+  } else {
+    psh <- psh %>% 
+      plotly::layout(showlegend = FALSE, scene=list(camera=.plotly3d$camera))
+    assign("plotlyscenehandle", psh, envir=.plotly3d)
+    psh
+  }
 }
 
 #' @rdname plot.neuron