+++ R/neuron-io.R
@@ -162,7 +162,7 @@ read.neuron<-function(f, format=NULL, class=c("neuron", "ngraph"), ...){
 #' # attach metadata to neuronlist
 #' vdnl=read.neurons(vdurls, neuronnames=vds, df=vddf)
 #' # use metadata to plot a subset of neurons
-#' clear3d()
+#' nclear3d()
 #' plot3d(vdnl, grepl("P[1-6].app", Lineage))
 #' }
 #' @importFrom tools file_path_sans_ext