+++ R/zzz.R
@@ -61,6 +61,11 @@ update_igraph <- FALSE
     update_igraph <- TRUE
   }
   invisible()
+  
+  # Set the default plot engine
+  if (is.null(getOption('nat.plotengine'))){
+    options(nat.plotengine='rgl')
+  }
 }
 
 .onAttach <- function(libname, pkgname) {
@@ -82,3 +87,9 @@ update_igraph <- FALSE
 
 # Will store stack of plotted rgl objects, ready for popping
 .plotted3d <- new.env()
+
+# Will store stack of plotted plotly objects, ready for popping
+.plotly3d <- new.env()
+.plotly3d$camera = list(up=list(x=0, y=0, z=1),
+              center=list(x=0, y=0, z=0),
+              eye=list(x=-0.1, y=-2.5, z=0.1))